Repository refactor
