+++ src/main/kotlin/edu/agh/susgame/dto/socket/common/GameStatus.kt
@@ -1,4 +1,4 @@
-package edu.agh.susgame.dto.common
+package edu.agh.susgame.dto.socket.common
 
 enum class GameStatus {
     WAITING,
