+++ src/main/kotlin/edu/agh/susgame/dto/socket/server/EdgeDTO.kt
@@ -1,4 +1,4 @@
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.socket.server
 
 import kotlinx.serialization.Serializable
